rename files to plural, add tags to create response
